Nest the Watch, Scalac, Signature and Image satellites
`WatchError`, `WatchEvent`, `WatchHandle`, `WatchOpenable` and `WatchAllOpenable`
become `Watch.*`; `ScalacSession` and `ScalacSessional` become `Scalac.*`;
`SignatureDigest` becomes `Signature.Digest`; `ImageDataOpenable` becomes
`Image.DataOpenable`. Nine names leave the umbrella's exported surface.

`Watch.Openable` and `Scalac.Sessional` shadowed the `aperture.Openable` and
`aperture.Sessional` they extend, both qualified rather than abandoned.

`TelReader` was attempted and reverted: nesting it makes `Tel`'s parse loop fail
with "cannot call update method entryOrdinal since its capture set {} is
read-only" — the read-only-versus-exclusive failure again, in a third shape. The
first two were `Array` indexing; this one is a `+=` on a mutable buffer, so the
signature is broader than indexing alone.

The sweep also rewrote `jnf.WatchEvent` — `java.nio.file`'s, not ours — which is
the foreign-type hazard the proposal warns about; caught by the compiler, but it
is why a bare-name sweep needs its output read.
