+++ lib/anthology/src/scala/anthology_scala.scala
@@ -203,11 +203,11 @@ private[anthology] def progressCallback(process: CompileProcess): dtdsi.Progress
 // Sugar for compiling within a session scope, `sources.compile()`, delegating to the
 // contextual session handle's own method.
 extension (sources: Map[Text, Text])
-  inline def compile()(using session: ScalacSession^)
-  :   ScalacSession.Process^{session, scala.caps.any} =
+  inline def compile()(using session: Scalac.Session^)
+  :   Scalac.Session.Process^{session, scala.caps.any} =
 
     session.compile(sources)
 
 // The contextual session handle, `transparent inline` so its precise, scoped type
 // survives the summon.
-transparent inline def compilation(using session: ScalacSession^): session.type = session
+transparent inline def compilation(using session: Scalac.Session^): session.type = session
