+++ lib/surveillance/src/core/soundness_surveillance_core.scala
@@ -32,7 +32,7 @@
                                                                                                   */
 package soundness
 
-export surveillance.{NativeWatcher, PollingWatcher, Watch, watch, WatchAllOpenable, Watcher, WatchError, WatchEvent, WatchHandle, WatchOpenable}
+export surveillance.{NativeWatcher, PollingWatcher, Watch, watch, Watcher}
 
 package watchers:
   export surveillance.watchers.{nativeWatcher, polling}
