+++ lib/surveillance/src/core/surveillance.Watcher.scala
@@ -54,5 +54,5 @@ object Watcher:
 trait Watcher:
   // Pure (`->`) filters: they are constructed by `Watch.apply` from path names alone, and
   // registrations are retained in backend-global state, which must not capture.
-  def watch(directories: Map[jnf.Path, Text -> Boolean], spool: Relay[WatchEvent])
-  :   Watcher.Registration raises WatchError
+  def watch(directories: Map[jnf.Path, Text -> Boolean], spool: Relay[Watch.Event])
+  :   Watcher.Registration raises Watch.Error
